fix(gameproto): correct VarInt range checks to allow signed values
The G115 bounds checks added earlier rejected negative VarInts, but the Minecraft protocol encodes signed 32-bit VarInts in two's complement, so legitimate values such as protocol version -1 were refused. Overflow is now detected on the fifth byte's data bits instead of by comparing the accumulated value, and the Terraria length comparison widens to int64 rather than narrowing to int32.

+++ gameproto/terraria.go
@@ -151,7 +151,7 @@ func readTerrariaString(r *bytes.Reader) (string, error) {
 	}
 
 	rlen := r.Len()
-	if rlen < 0 || length > int32(rlen) {
+	if rlen < 0 || int64(length) > int64(rlen) {
 		return "", fmt.Errorf("string length %d exceeds remaining %d", length, rlen)
 	}
